test(infinite-scroll): migrate basic
